requirements.txt
# 基础依赖
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0

# 文档处理
python-docx>=0.8.11
openpyxl>=3.1.0

# 数据处理和可视化
plotly>=5.15.0
altair>=5.0.0

# 机器学习和向量处理（可选，用于真实RAG实现）
scikit-learn>=1.3.0
scipy>=1.11.0

# 文本处理
jieba>=0.42.1

# DeepSeek API 相关依赖
openai>=1.0.0
urllib3>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

# 工具库
tqdm>=4.65.0
requests>=2.31.0